Former-commit-id: 8f9c9269e0733dfe5a26d0e4990b5213e495eaaa
Former-commit-id: ab55462e2b887f9df4cf3c10628255ce52ffc375

+++ pre-registration/pre-registration-batchjob-service/src/main/java/io/mosip/preregistration/batchjobservices/service/ConsumedStatusService.java
@@ -11,6 +11,7 @@ import java.util.List;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -52,6 +53,12 @@ public class ConsumedStatusService {
 	
 	@Value("${mosip.utc-datetime-pattern}")
 	private String utcDateTimePattern;
+	
+	@Value("${ver}")
+	String versionUrl;
+
+	@Value("${id}")
+	String idUrl;
 
 	/**
 	 * Autowired reference for {@link #batchServiceDAO}
@@ -65,6 +72,7 @@ public class ConsumedStatusService {
 	 * 
 	 * @return Response DTO
 	 */
+	
 	@Transactional(propagation = Propagation.REQUIRES_NEW, rollbackFor = Exception.class)
 	public MainResponseDTO<String> demographicConsumedStatus() {
 
@@ -89,9 +97,11 @@ public class ConsumedStatusService {
 					batchServiceDAO.updateConsumedDemographic(demographicEntityConsumed);
 
 					DocumentEntity documentEntity = batchServiceDAO.getDocumentDetails(preRegId);
-					BeanUtils.copyProperties(documentEntity, documentEntityConsumed);
-					batchServiceDAO.updateConsumedDocument(documentEntityConsumed);
-
+					if(documentEntity!=null) {
+						BeanUtils.copyProperties(documentEntity, documentEntityConsumed);
+						batchServiceDAO.updateConsumedDocument(documentEntityConsumed);
+					}
+					
 					RegistrationBookingEntity bookingEntity = batchServiceDAO.getPreRegId(preRegId);
 					BeanUtils.copyProperties(bookingEntity, bookingEntityConsumed);
 					RegistrationBookingPKConsumed consumedPk=new RegistrationBookingPKConsumed();
@@ -100,23 +110,26 @@ public class ConsumedStatusService {
 					bookingEntityConsumed.setBookingPK(consumedPk);
 					batchServiceDAO.updateConsumedBooking(bookingEntityConsumed);
 
+					if(documentEntity!=null) {
+						batchServiceDAO.deleteDocument(documentEntity);
+					}
 					batchServiceDAO.deleteBooking(bookingEntity);
-					batchServiceDAO.deleteDocument(documentEntity);
 					batchServiceDAO.deleteDemographic(demographicEntity);
-					log.info("sessionId", "idType", "id", "Update the status successfully into Consumed tables Pre-Registration");
+					log.info("sessionId", "idType", "id", "Update the status successfully into Consumed tables for Pre-RegistrationId: "+preRegId);
+					
+					iterate.setStatusComments(NEW_STATUS_COMMENTS);
+					batchServiceDAO.updateProcessedList(iterate);
+					log.info("sessionId", "idType", "id", "Update the comment successfully into Processed PreId List table for Pre-RegistrationId: "+preRegId);
 				}
-
-				iterate.setStatusComments(NEW_STATUS_COMMENTS);
-				batchServiceDAO.updateProcessedList(iterate);
-				log.info("sessionId", "idType", "id", "Update the comment successfully into Processed PreId List table");
 
 			});
 
 		} catch (Exception e) {
 			new BatchServiceExceptionCatcher().handle(e);
 		}
-		response.setResTime(getCurrentResponseTime());
-		response.setStatus(true);
+		response.setResponsetime(getCurrentResponseTime());
+		response.setId(idUrl);
+		response.setVersion(versionUrl);
 		response.setResponse("Demographic status to consumed updated successfully");
 		return response;
 	}
